add logic to handle encrypted assertions
Current code wasn't handling encrypted assertions at all. Now they'll be handled as a fallback to standard assertions.

This involved taking a private key and passing it through. I added this in a way that is very backwards-compatible. I haven't added unit tests or even validated that they run- I was unable to set up a local environment that could run them.