add encrypted assertion support and test cases
This change, based on a patch originally contributed by Ted Timmons,
adds handling of encrypted assertions.  To use this, the service
provider provisions a keypair and gives the IdP the pubkey, which
the IdP uses to encrypt a symmetric key and the assertion with that.

The SP will need to set the private key in the SPConfig object
prior to processing encrypted assertions with the SAMLClient class.